Xử lý phần giỏ hàng

+++ src/screens/cart/cartItem/cart-item.component.js
@@ -1,26 +1,67 @@
-import { View, Text, Image, TouchableOpacity } from 'react-native'
-import React, { useState } from 'react'
+import { View, Text, Image, TouchableOpacity, Alert } from 'react-native'
+import React, { useEffect, useRef, useState } from 'react'
 import styles from './cart-item.styles'
 import CounterButton from '../../../components/counterButton/counter-button.component'
 import CheckBox from '../../../components/checkbox/checkbox.component'
 import { formatCurrency } from '../../../helpers/Utils'
 import { useDispatch } from 'react-redux'
-import { removeCartThunk } from '../../../redux/cart/cart.slice'
+import { calculateTotal, removeCartThunk, saveCheckStatusThunk } from '../../../redux/cart/cart.slice'
 
 const CartItem = ({ item }) => {
   const dispatch = useDispatch()
   const [quantity, setQuantity] = useState(item.quantity)
+  const checkBoxRef = useRef();
+
+  useEffect(() => {
+    const checked = checkBoxRef.current.isChecked()
+    checked && dispatch(calculateTotal({ checked, productPriceTotal: item.price * quantity }))
+  }, [])
+
   const removeProductInCart = () => {
-    dispatch(removeCartThunk({ productId: item.id }))
+    const checked = checkBoxRef.current.isChecked()
+    Alert.alert('Thông báo', 'Bạn có muốn xóa sản phẩm này không?', [
+      {
+        text: 'OK', onPress: () => {
+          dispatch(removeCartThunk({ productId: item.id }))
+          checked && dispatch(calculateTotal({ productPriceTotal: item.price * quantity, isDecrease: true }))
+        }
+      },
+      {
+        text: 'Hủy',
+        onPress: () => { },
+      },
+    ]);
   }
 
   const handleDecreaseQuantity = () => {
-    setQuantity(quantity - 1)
+    setQuantity((prevQuantity) => {
+      const newQuantity = prevQuantity - 1
+      if (checkBoxRef.current.isChecked()) {
+        dispatch(calculateTotal({ productPriceTotal: item.price * 1, isDecrease: true }))
+      }
+
+      if (newQuantity < 1) return 1
+      return newQuantity
+    })
   }
 
   const handleIncreaseQuantity = () => {
-    setQuantity(quantity + 1)
+    setQuantity((prevQuantity) => {
+      const newQuantity = prevQuantity + 1
+
+      if (checkBoxRef.current.isChecked()) {
+        dispatch(calculateTotal({ productPriceTotal: item.price * 1, isIncrease: true, }))
+      }
+
+      return newQuantity
+    })
   }
+
+  const onPress = (checked) => {
+    dispatch(calculateTotal({ checked, productPriceTotal: item.price * quantity }))
+    dispatch(saveCheckStatusThunk({ productId: item.id, checkStatus: checked }))
+  }
+
   return (
     <View style={styles.item}>
       <Image source={{ uri: item.image }} style={styles.image} />
@@ -43,6 +84,9 @@ const CartItem = ({ item }) => {
         <Text>{formatCurrency(item.price)}</Text>
 
         <CheckBox
+          ref={checkBoxRef}
+          onPress={onPress}
+          initialCheckStatus={item.check}
         />
       </View>
     </View>
